Give docker more time to allocate host ports
As it turns out, sometimes docker doesn't allocate ports on the host
immediately after the container is created and started. Sometimes the
first attempt to inspect a container will not include any ports bound on
the host. To tackle that issue, gnomock will retry container inspection
until it finds out enough allocated ports, or until the attempt times
out.